Move distribution methods to sub-submodule
